Fix due date filter: type export, code duplication, and backend guard
- Move DueDateOption/DueDateFilterValue types and computeDueDateFilters
  to shared utility (src/lib/utils/dueDateFilter.ts) to fix Svelte
  "export type" compile error and eliminate duplicated logic between
  DueDateFilter.svelte and +page.svelte
- Add mutual exclusion in backend: no_due_date filter now prevents
  start_date/end_date from being applied simultaneously
- Simplify reactive statement to clearly establish dependencies on
  selectedProjectId and selectedDueDate via local variable references
