+++ services/frontend/src/lib/components/DueDateFilter.svelte
@@ -1,58 +1,17 @@
 <script lang="ts">
 	import { createEventDispatcher } from 'svelte';
-	import { formatDateForInput, getStartOfWeek } from '$lib/utils/dates';
-
-	export type DueDateOption = 'all' | 'today' | 'this_week' | 'next_two_weeks' | 'no_due_date';
-
-	export interface DueDateFilterValue {
-		option: DueDateOption;
-		start_date?: string;
-		end_date?: string;
-		no_due_date?: boolean;
-	}
+	import { computeDueDateFilters } from '$lib/utils/dueDateFilter';
+	import type { DueDateOption, DueDateFilterValue } from '$lib/utils/dueDateFilter';
 
 	export let selected: DueDateOption = 'all';
 
 	const dispatch = createEventDispatcher<{ change: DueDateFilterValue }>();
 
-	function getFilterValue(option: DueDateOption): DueDateFilterValue {
-		const today = new Date();
-		const todayStr = formatDateForInput(today);
-
-		switch (option) {
-			case 'today':
-				return { option, start_date: todayStr, end_date: todayStr };
-			case 'this_week': {
-				const weekStart = getStartOfWeek(today);
-				const weekEnd = new Date(weekStart);
-				weekEnd.setDate(weekEnd.getDate() + 6);
-				return {
-					option,
-					start_date: formatDateForInput(weekStart),
-					end_date: formatDateForInput(weekEnd)
-				};
-			}
-			case 'next_two_weeks': {
-				const twoWeeksEnd = new Date(today);
-				twoWeeksEnd.setDate(twoWeeksEnd.getDate() + 14);
-				return {
-					option,
-					start_date: todayStr,
-					end_date: formatDateForInput(twoWeeksEnd)
-				};
-			}
-			case 'no_due_date':
-				return { option, no_due_date: true };
-			default:
-				return { option: 'all' };
-		}
-	}
-
 	function handleChange(event: Event) {
 		const target = event.target as HTMLSelectElement;
 		const option = target.value as DueDateOption;
 		selected = option;
-		dispatch('change', getFilterValue(option));
+		dispatch('change', { option, ...computeDueDateFilters(option) });
 	}
 </script>
